+++ R/faux_na_if.r
@@ -3,20 +3,18 @@
 #'     is_atomic is_formula is_function is_list is_logical is_missing warn
 
 faux_na_if <- function(
-  input, ..., arguments, arg_names, not = FALSE, scoped = FALSE
+  x, ..., arguments, arg_names, not = FALSE, scoped = FALSE
 ) {
   if (!scoped) {
     arguments <- list(...)
     arg_names <- as.list(substitute(list(...)))
   }
 
-  original_input <- input
-
-  if (is_list(recurse(input))) abort_uncoercible(input)
+  if (is_list(recurse(x))) abort_uncoercible(x)
 
   if (length(arguments) == 0) {
     warn("No values to replace with `NA` specified.")
-    return(input)
+    return(x)
   }
 
   lists <- vapply(arguments, is_list, logical(1))
@@ -32,20 +30,20 @@ faux_na_if <- function(
   eval_replacement <- function(replacement) {
     if (is_atomic(replacement)) return(replacement)
 
-    replacement <- as_function(replacement)(input)
+    replacement <- as_function(replacement)(x)
 
     if (!is_logical(replacement)) return(replacement)
 
-    input[replacement]
+    x[replacement]
   }
 
   replacements <- recurse(lapply(arguments[valid], eval_replacement))
 
   if (length(replacements) > 0) {
     if (not) {
-      input[!input %in% replacements] <- NA
+      x[!x %in% replacements] <- NA
     } else {
-      input[input %in% replacements] <- NA
+      x[x %in% replacements] <- NA
     }
   }
 
@@ -54,7 +52,7 @@ faux_na_if <- function(
   if (any(two_sided))           warn_two_sided(arg_names[-1][two_sided])
   if (any(!valid & !two_sided)) warn_invalid(arg_names[-1][!valid & !two_sided])
 
-  input
+  x
 }
 
 scoped_na_if <- function(fun, .tbl, ...) {