+++ R/fauxnaif-package.R
@@ -4,6 +4,6 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
-#' @importFrom lifecycle deprecate_soft
+#' @importFrom lifecycle deprecate_warn
 ## usethis namespace: end
 NULL