+++ R/scoped_na_if.r
@@ -51,9 +51,7 @@ NULL
 #' @export
 
 na_if_all <- function(.tbl, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun       = dplyr::mutate_all,
@@ -67,9 +65,7 @@ na_if_all <- function(.tbl, ...) {
 #' @export
 
 na_if_not_all <- function(.tbl, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun       = dplyr::mutate_all,
@@ -84,9 +80,7 @@ na_if_not_all <- function(.tbl, ...) {
 #' @export
 
 na_if_at <- function(.tbl, .vars, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun       = dplyr::mutate_at,
@@ -101,9 +95,7 @@ na_if_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_not_at <- function(.tbl, .vars, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun       = dplyr::mutate_at,
@@ -119,9 +111,7 @@ na_if_not_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_if <- function(.tbl, .predicate, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun        = dplyr::mutate_if,
@@ -136,9 +126,7 @@ na_if_if <- function(.tbl, .predicate, ...) {
 #' @export
 
 na_if_not_if <- function(.tbl, .predicate, ...) {
-  if (!identical(Sys.getenv("TESTTHAT"), "true")) {
-    lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
-  }
+  lifecycle::deprecate_warn("0.6.0", format(sys.call()[1]), "dplyr::across()")
 
   scoped_na_if(
     fun        = dplyr::mutate_if,