+++ NEWS.md
@@ -1,5 +1,8 @@
 # fauxnaif (development version)
 
+* Replace argument `input` with `x`.
+* Advance `fauxnaif::na_if()` and scoped `na_if_*()` functions from `deprecate_soft()` to `deprecate_warn()`.
+* Only warn `fauxnaif::na_if()` is deprecated if arguments could not be handled by `dplyr::na_if()`.
 * Remove `inst/CITATION`.
 
 # fauxnaif 0.6.0